pyproject.toml
[tool.poetry]
name = "col"
version = "0.1.0"
description = "Package for creating attendence certificates"
authors = ["Zak Oster"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pypdf = "^3.11.1"
pandas = "^2.0.3"
openpyxl = "^3.1.2"
pillow = "^10.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
black = "^23.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
